[FrameworkBundle][Routing][Translation][TwigBundle] Move the last classes the components name into their own packages
