+++ src/Symfony/Bundle/FrameworkBundle/Routing/Attribute/AsRoutingConditionService.php
@@ -11,44 +11,8 @@
 
 namespace Symfony\Bundle\FrameworkBundle\Routing\Attribute;
 
-use Symfony\Component\DependencyInjection\Attribute\AutoconfigureTag;
+use Symfony\Component\Routing\Attribute\AsRoutingConditionService as BaseAsRoutingConditionService;
 
-/**
- * Service tag to autoconfigure routing condition services.
- *
- * You can tag a service:
- *
- *     #[AsRoutingConditionService('foo')]
- *     class SomeFooService
- *     {
- *         public function bar(): bool
- *         {
- *             // ...
- *         }
- *     }
- *
- * Then you can use the tagged service in the routing condition:
- *
- *     class PageController
- *     {
- *         #[Route('/page', condition: "service('foo').bar()")]
- *         public function page(): Response
- *         {
- *             // ...
- *         }
- *     }
- */
-#[\Attribute(\Attribute::TARGET_CLASS)]
-class AsRoutingConditionService extends AutoconfigureTag
-{
-    /**
-     * @param string|null $alias    The alias of the service to use it in routing condition expressions
-     * @param int         $priority Defines a priority that allows the routing condition service to override a service with the same alias
-     */
-    public function __construct(
-        ?string $alias = null,
-        int $priority = 0,
-    ) {
-        parent::__construct('routing.condition_service', ['alias' => $alias, 'priority' => $priority]);
-    }
-}
+trigger_deprecation('symfony/framework-bundle', '8.2', 'The "%s" class is deprecated, use "%s" instead.', 'Symfony\Bundle\FrameworkBundle\Routing\Attribute\AsRoutingConditionService', BaseAsRoutingConditionService::class);
+
+class_alias(BaseAsRoutingConditionService::class, 'Symfony\Bundle\FrameworkBundle\Routing\Attribute\AsRoutingConditionService');
