+++ src/Symfony/Bundle/TwigBundle/Resources/config/twig.php
@@ -34,6 +34,7 @@ use Symfony\Bridge\Twig\Extension\WorkflowExtension;
 use Symfony\Bridge\Twig\Extension\YamlExtension;
 use Symfony\Bridge\Twig\Translation\TwigExtractor;
 use Symfony\Bundle\TwigBundle\CacheWarmer\TemplateCacheWarmer;
+use Symfony\Bundle\TwigBundle\Controller\TemplateController;
 use Symfony\Bundle\TwigBundle\DependencyInjection\Configurator\EnvironmentConfigurator;
 use Symfony\Bundle\TwigBundle\TemplateIterator;
 use Twig\Cache\ChainCache;
@@ -198,5 +199,16 @@ return static function (ContainerConfigurator $container) {
         ->set('controller.template_attribute_listener', TemplateAttributeListener::class)
             ->args([service('twig')])
             ->tag('kernel.event_subscriber')
+
+        ->set(TemplateController::class)
+            ->args([
+                service('twig'),
+            ])
+            ->public()
+
+        // the controller used to be declared by the routing configuration under its old name,
+        // which applications reference from their route definitions
+        ->alias('Symfony\\Bundle\\FrameworkBundle\\Controller\\TemplateController', TemplateController::class)
+            ->public()
     ;
 };
