+++ src/Symfony/Component/Routing/CHANGELOG.md
@@ -5,6 +5,8 @@ CHANGELOG
 ---
 
  * Add `RouterBundle`, which provides the `router` configuration and the services previously provided by `FrameworkBundle` under `framework.router`
+ * Add the `#[AsRouteLoader]` attribute, which replaces `FrameworkBundle`'s `RouteLoaderInterface`
+ * Add `DependencyInjection\Router` and `Attribute\AsRoutingConditionService`, which `FrameworkBundle` used to provide
  * Add `CacheWarmer\RouterCacheWarmer`, `Controller\RedirectController`, `Loader\AttributeRouteControllerLoader`, `Loader\DelegatingLoader` and `Matcher\RedirectableCompiledUrlMatcher`, which `FrameworkBundle` used to provide
  * Add a `firewall` option to route definitions, exposed as the `_firewall` route default
  * Allow defining default query parameters with the `_query` route default
