+++ src/Symfony/Component/Translation/Tests/Fixtures/Resources/translations/domain.with.dots.en.yml
