+++ src/Symfony/Component/Translation/Tests/Fixtures/Resources/translations/messages.fr.yml
