+++ src/Symfony/Component/Translation/Tests/Fixtures/Resources/translations2/ccc.fr.yml
